test case with some notes on serializing to SMT2
Currently there is not an obvious way to serialize to a complete
SMT2 file using the pysmt API *but* there is a way to do so with
the Z3 solver implementation.

Ideally there would be a generic solution but it will require a
slightly deeper dive into the documentation and/or code so for now
this commit holds a sample of the Z3 solution.